requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.11.5
httpx==0.27.0
PyJWT==2.10.1
